Add windows tests
